refactor(JobReducers): extract specialized reducers
Extract speclialized reducers from JobReducers to separate files, to prevent the JobReducers file to balloon in size/complexity over time.
